+++ src/itba/edu/ar/cellIndexMethod/route/Route.java
@@ -8,7 +8,7 @@ import itba.edu.ar.cellIndexMethod.data.particle.Particle;
 
 public interface Route {
 
-	public void fillNeightbours(int x,int y,IndexMatrix matrix,Particle particle,float interactionRadio);
+	public void fillNeightbours(int x,int y,IndexMatrix matrix,Particle particle,double interactionRadio);
 
 	
 }
